fix : corrections suite au merge

+++ src/Vue/Administration/vueDetailEntreprise.php
@@ -15,10 +15,12 @@ if(isset($_GET['numSiret'])){
     echo "<h1>".htmlspecialchars($entreprise->getNomEntreprise())."</h1>";
     echo "<h2> Numéro SIRET : " . htmlspecialchars($entreprise->getNumSiret()) . "</h2>";
 
-    echo '<div class="boutonsGeneral">';
-    echo '<a  href="controleurFrontal.php?action=afficherMAJEntreprise&numSiret='.$entreprise->getNumSiret().'"> Modifier les informations de l\'Entreprise </a>';
-    echo '<a  href="controleurFrontal.php?action=afficherDeleteEntreprise&numSiret='.$entreprise->getNumSiret().'"> Supprimer le compte de l\'Entreprise </a>';
-    echo '</div>';
+    if(\App\Lib\ConnexionUtilisateur::estMaitreSA()){
+        echo '<div class="boutonsGeneral">';
+        echo '<a  href="controleurFrontal.php?action=afficherMAJEntreprise&numSiret='.$entreprise->getNumSiret().'"> Modifier les informations de l\'Entreprise </a>';
+        echo '<a  href="controleurFrontal.php?action=afficherDeleteEntreprise&numSiret='.$entreprise->getNumSiret().'"> Supprimer le compte de l\'Entreprise </a>';
+        echo '</div>';
+    }
 
     echo '<hr/>';
